backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.8.1
asttokens==2.4.0
attrs==25.3.0
Authlib==1.3.1
backcall==0.2.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
cryptography==45.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
deepseek==1.0.0
deprecation==2.1.0
distro==1.9.0
Django==3.1.12
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.7
dnspython==2.7.0
email_validator==2.2.0
executing==1.2.0
fastapi==0.104.1
fastapi-cli==0.0.7
fonttools==4.43.1
greenlet==3.2.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.26.0
openai==1.82.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==5.29.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==3.11.4
pyOpenSSL==24.0.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.20
pytz==2023.3.post1
PyYAML==6.0.2
pyzmq==25.1.1
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.6
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
selenium==4.15.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.2.4
stack-data==0.6.2
starlette==0.27.0
threadpoolctl==3.2.0
tornado==6.3.3
tqdm==4.67.1
traitlets==5.10.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.3
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.24.0
uvloop==0.21.0
validators==0.34.0
watchfiles==1.0.5
wcwidth==0.2.6
weaviate-client==4.14.1
webdriver-manager==4.0.1
websockets==15.0.1
wsproto==1.2.0
